move variant fetcher to module
