+++ app/controllers/variants_controller.rb
@@ -1,26 +1,9 @@
 # frozen_string_literal: true
 
 class VariantsController < ApplicationController
+  include VariantsFetcher
 
   def index
     render json: {variants: variants}
-  end
-  
-  def variants
-    filter_id = params["filter_id"];
-    if filter_id.empty?
-      adapter.variants
-    else
-      adapter.variants_in_filter(filter_id)
-    end
-  end
-
-  def adapter
-    channel_id = params["channel_id"];
-    if (channel_id || '').empty?
-      VariantFetcher.new(current_user)
-    else
-      ChannelVariantFetcher.new(current_user,channel_id)
-    end
   end
 end
